🐛 [Fix] 이미지가 업로드되지 않았던 버그 수정
- 서버 통신 모델 타입의 price와 discountedPrice 타입을 Int로 변경
- parameter를 만드는 방식을 asDictionary() -> Mirror로 변경
- parameter의 value가 nil인 경우 body에 넣지 않도록 변경
- 테스트를 위해 ItemList 타입에 Equatable 프로토콜 채택
- 테스트를 위해 ItemResponse 타입을 Codable로 변경하고 Equatable 프로토콜 채택
- Asset에 테스트 image 추가

+++ OpenMarket/OpenMarket/Model/Request/ItemForEdit.swift
@@ -9,11 +9,11 @@ import Foundation
 
 struct ItemForEdit: Codable {
     let title: String?
-    let price: Double?
+    let price: Int?
     let descriptions: String?
     let currency: String?
     let stock: Int?
-    let discountedPrice: Double?
+    let discountedPrice: Int?
     let images: [Data]?
     let password: String
